requirements.txt
requests
python-dotenv
pyaudio
